Platform-specific starvation checker tests
