+++ tests/jvm/src/main/scala/catseffect/examplesplatform.scala
@@ -17,10 +17,12 @@
 package catseffect
 
 import cats.effect.{ExitCode, IO, IOApp}
+import cats.effect.unsafe.IORuntimeConfig
+import cats.effect.std.Random
 import cats.syntax.all._
 
 import scala.concurrent.ExecutionContext
-import scala.concurrent.duration.Duration
+import scala.concurrent.duration._
 
 import java.io.File
 
@@ -75,4 +77,27 @@ package examples {
         case _ => ExitCode.Error
       }
   }
+
+  // The parameters here were chosen experimentally and seem to be
+  // relatively reliable. The trick is finding a balance such that
+  // we end up with every WSTP thread being blocked but not permanently
+  // so that the starvation detector fiber still gets to run and
+  // therefore report its warning
+  object CpuStarvation extends IOApp.Simple {
+
+    override protected def runtimeConfig: IORuntimeConfig = IORuntimeConfig().copy(
+      cpuStarvationCheckInterval = 200.millis,
+      cpuStarvationCheckInitialDelay = 0.millis,
+      cpuStarvationCheckThreshold = 0.2d
+    )
+
+    val run = Random.scalaUtilRandom[IO].flatMap { rand =>
+      // jitter to give the cpu starvation checker a chance to run at all
+      val jitter = rand.nextIntBounded(100).flatMap(n => IO.sleep(n.millis))
+      (jitter >> IO(Thread.sleep(400)))
+        .replicateA_(10)
+        .parReplicateA_(Runtime.getRuntime().availableProcessors() * 2)
+    }
+  }
+
 }
